Added some explicit type conversions.
git-svn-id: svn+ssh://svn.code.sf.net/p/hfst/code/trunk/hfst3@4312 941e2c2b-deac-454f-805a-451daa25f33c

+++ libhfst/src/implementations/LogWeightTransducer.cc
@@ -1691,12 +1691,13 @@ namespace hfst { namespace implementations
   {
     assert(t->InputSymbols() != NULL);
     SymbolTable * st = t->InputSymbols()->Copy();
+    // TODO: cl.exe: conversion from 'int64' to 'unsigned int'
     pair<unsigned int, unsigned int> old_pair
-      (st->AddSymbol(old_symbol_pair.first),
-       st->AddSymbol(old_symbol_pair.second));
+      ((unsigned int)st->AddSymbol(old_symbol_pair.first),
+       (unsigned int)st->AddSymbol(old_symbol_pair.second));
     pair<unsigned int, unsigned int> new_pair
-      (st->AddSymbol(new_symbol_pair.first),
-       st->AddSymbol(new_symbol_pair.second));
+      ((unsigned int)st->AddSymbol(new_symbol_pair.first),
+       (unsigned int)st->AddSymbol(new_symbol_pair.second));
     LogFst * retval = substitute(t, old_pair, new_pair);
     retval->SetInputSymbols(st);
     delete st;
